separate stock component to add and change
